Fix my mail signature command again
If I define it within the use-package call for Markdown Mode then it
doesn’t have any effect.  And my previous idea to do it in Mail Mode
was stupid because it has no access to the ‘markdown-mode-map’ table.
This patch defines my signature function outside the scope of all
packages so that I can bind it within the setup for Markdown Mode.